requirements.txt
carla==0.9.16
numpy>=1.24
Pillow>=10.0
opencv-python>=4.8
matplotlib>=3.8
pyyaml>=6.0

# For CARLA BasicAgent (third_party/carla_pythonapi)
networkx
shapely